[not verified] move url helpser to client/ level
