+++ projects/packages/videopress/src/client/block-editor/blocks/video/save.js
@@ -6,7 +6,7 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import { getVideoPressUrl } from '../../utils/url/index.js';
+import { getVideoPressUrl } from '../../../utils/url/index.js';
 
 /**
  * VideoPress block save function
