+++ projects/packages/videopress/src/client/block-editor/extend/core-embed/edit.js
@@ -13,7 +13,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import './editor.scss';
-import { pickGUIDFromUrl } from '../../utils/url/index.js';
+import { pickGUIDFromUrl } from '../../../utils/url/index.js';
 
 const withCoreEmbedVideoPressBlock = createHigherOrderComponent( CoreEmbedBlockEdit => {
 	return props => {
